Presentation: Filtering APIs docs (#4959)
* Document base filter builder component

* Cleanup API

* Document filtering APIs in presentation-components

* Group PropertyFilterBuilder docs

* Add NextVersion entry

* Update docs/changehistory/NextVersion.md


* Update docs/changehistory/NextVersion.md


* Update docs/changehistory/NextVersion.md


* Update docs/changehistory/NextVersion.md


* Update presentation/components/src/presentation-components/instance-filter-builder/PresentationInstanceFilterProperty.tsx


* Update presentation/components/src/presentation-components/instance-filter-builder/PresentationInstanceFilterProperty.tsx


* Update presentation/components/src/presentation-components/tree/controlled/PresentationTreeNodeRenderer.tsx


* Update presentation/components/src/presentation-components/tree/controlled/PresentationTreeRenderer.tsx


* Clean up docs

* Make some APIs internal

* extract-api

* rush change

* Fix reference link

* Rename and context value to have different names
