+++ common/api/summary/presentation-components.exports.csv
@@ -6,12 +6,9 @@ public;CacheInvalidationProps
 public;ContentDataProvider 
 public;ContentDataProviderProps 
 public;ControlledPresentationTreeFilteringProps
-alpha;convertPresentationFilterToPropertyFilter(descriptor: Descriptor, filter: PresentationInstanceFilter): PropertyFilter | undefined
-alpha;convertToInstanceFilterDefinition(filter: PresentationInstanceFilter, imodel: IModelConnection): Promise
+beta;convertToInstanceFilterDefinition(filter: PresentationInstanceFilter, imodel: IModelConnection): Promise
 alpha;createDiagnosticsOptions(props: DiagnosticsProps): ClientDiagnosticsOptions | undefined
 internal;createFieldInfo(field: Field, parentFieldName?: string):
-alpha;createInstanceFilterPropertyInfos(descriptor: Descriptor): InstanceFilterPropertyInfo[]
-internal;createPresentationInstanceFilter(descriptor: Descriptor, filter: PropertyFilter): PresentationInstanceFilter | undefined
 internal;createPropertyDescriptionFromFieldInfo(info: FieldInfo): PropertyDescription
 public;DataProvidersFactory
 deprecated;DataProvidersFactory
@@ -30,14 +27,9 @@ internal;FieldRecord
 internal;FilteredPresentationTreeDataProvider 
 internal;FilteredPresentationTreeDataProviderProps
 internal;getFavoritesCategory: () => CategoryDescription
-internal;getInstanceFilterFieldName(property: PropertyDescription): string
 public;IContentDataProvider 
 public;IFavoritePropertiesDataProvider
 public;IFilteredPresentationTreeDataProvider 
-alpha;INSTANCE_FILTER_FIELD_SEPARATOR = "#"
-alpha;InstanceFilterBuilder(props: InstanceFilterBuilderProps): JSX.Element
-alpha;InstanceFilterBuilderProps
-alpha;InstanceFilterPropertyInfo
 beta;InstanceKeyValueRenderer 
 public;IPresentationDataProvider 
 public;IPresentationLabelsProvider
@@ -46,26 +38,24 @@ public;IPresentationTableDataProvider = TableDataProvider & IContentDataProvider
 deprecated;IPresentationTableDataProvider = TableDataProvider & IContentDataProvider &
 public;IPresentationTreeDataProvider 
 internal;IPropertiesAppender
-alpha;isPresentationInfoTreeNodeItem(item: TreeNodeItem): item is PresentationInfoTreeNodeItem
-alpha;isPresentationInstanceFilterConditionGroup(filter: PresentationInstanceFilter): filter is PresentationInstanceFilterConditionGroup
-alpha;isPresentationTreeNodeItem(item: TreeNodeItem): item is PresentationTreeNodeItem
+beta;isPresentationInfoTreeNodeItem(item: TreeNodeItem): item is PresentationInfoTreeNodeItem
+beta;isPresentationInstanceFilterConditionGroup(filter: PresentationInstanceFilter): filter is PresentationInstanceFilterConditionGroup
+beta;isPresentationTreeNodeItem(item: TreeNodeItem): item is PresentationTreeNodeItem
 public;IUnifiedSelectionComponent
-alpha;NavigationPropertyEditor 
-alpha;NavigationPropertyEditorContext
-alpha;navigationPropertyEditorContext: React_2.Context
-alpha;NavigationPropertyTargetEditor 
+beta;NavigationPropertyEditor 
+beta;navigationPropertyEditorContext: React_2.Context
+beta;NavigationPropertyEditorContextProps
+beta;NavigationPropertyTargetEditor 
 public;PresentationComponentsLoggerCategory
-alpha;PresentationInfoTreeNodeItem 
-alpha;PresentationInstanceFilter = PresentationInstanceFilterConditionGroup | PresentationInstanceFilterCondition
-alpha;PresentationInstanceFilterBuilder(props: PresentationInstanceFilterBuilderProps): JSX.Element
-alpha;PresentationInstanceFilterBuilderProps
-alpha;PresentationInstanceFilterCondition
-alpha;PresentationInstanceFilterConditionGroup
-alpha;PresentationInstanceFilterDialog(props: PresentationInstanceFilterDialogProps): JSX.Element
-alpha;PresentationInstanceFilterDialogProps 
-alpha;PresentationInstanceFilterInfo
-alpha;PresentationInstanceFilterProperty(props: PresentationInstanceFilterPropertyProps): JSX.Element
-alpha;PresentationInstanceFilterPropertyProps
+beta;PresentationInfoTreeNodeItem 
+beta;PresentationInstanceFilter = PresentationInstanceFilterConditionGroup | PresentationInstanceFilterCondition
+beta;PresentationInstanceFilterBuilder(props: PresentationInstanceFilterBuilderProps): JSX.Element
+beta;PresentationInstanceFilterBuilderProps
+beta;PresentationInstanceFilterCondition
+beta;PresentationInstanceFilterConditionGroup
+beta;PresentationInstanceFilterDialog(props: PresentationInstanceFilterDialogProps): JSX.Element
+beta;PresentationInstanceFilterDialogProps 
+beta;PresentationInstanceFilterInfo
 public;PresentationLabelsProvider 
 public;PresentationLabelsProviderProps
 public;PresentationPropertyDataProvider 
@@ -77,14 +67,14 @@ deprecated;PresentationTableDataProviderProps
 public;PresentationTreeDataProvider 
 beta;PresentationTreeDataProviderDataSourceEntryPoints
 public;PresentationTreeDataProviderProps 
-alpha;PresentationTreeNodeItem 
-alpha;PresentationTreeNodeItemFilteringInfo
+beta;PresentationTreeNodeItem 
+beta;PresentationTreeNodeItemFilteringInfo
 public;PresentationTreeNodeLoaderProps 
 public;PresentationTreeNodeLoaderResult
-alpha;PresentationTreeNodeRenderer(props: PresentationTreeNodeRendererProps): JSX.Element
-alpha;PresentationTreeNodeRendererProps 
-alpha;PresentationTreeRenderer(props: PresentationTreeRendererProps): JSX.Element
-alpha;PresentationTreeRendererProps 
+beta;PresentationTreeNodeRenderer(props: PresentationTreeNodeRendererProps): JSX.Element
+beta;PresentationTreeNodeRendererProps 
+beta;PresentationTreeRenderer(props: PresentationTreeRendererProps): JSX.Element
+beta;PresentationTreeRendererProps 
 public;PropertyDataProviderWithUnifiedSelectionProps
 internal;class PropertyRecordsBuilder 
 internal;ReloadedHierarchyPart
@@ -104,13 +94,11 @@ beta;UnifiedSelectionState = (selectionLevel?: number) => Readonly
 public;UnifiedSelectionTreeEventHandler 
 public;UnifiedSelectionTreeEventHandlerParams
 public;useControlledPresentationTreeFiltering(props: ControlledPresentationTreeFilteringProps):
-internal;useFilterBuilderNavigationPropertyEditorContext(imodel: IModelConnection, descriptor: Descriptor): NavigationPropertyEditorContext
 internal;useFilteredNodeLoader(nodeLoader: AbstractTreeNodeLoaderWithProvider
-alpha;useHierarchyLevelFiltering(props: UseHierarchyLevelFilteringProps):
-alpha;UseHierarchyLevelFilteringProps
-alpha;useNavigationPropertyEditingContext(imodel: IModelConnection, dataProvider: IContentDataProvider): NavigationPropertyEditorContext
+beta;useHierarchyLevelFiltering(props: UseHierarchyLevelFilteringProps):
+beta;UseHierarchyLevelFilteringProps
+beta;useNavigationPropertyEditingContext(imodel: IModelConnection, dataProvider: IContentDataProvider): NavigationPropertyEditorContextProps
 internal;useNodeHighlightingProps(filter: string | undefined, filteredNodeLoader?: ITreeNodeLoaderWithProvider
-alpha;usePresentationInstanceFilteringProps(descriptor: Descriptor, imodel: IModelConnection, initialClasses?: ClassInfo[]):
 public;usePresentationTreeNodeLoader(props: PresentationTreeNodeLoaderProps): PresentationTreeNodeLoaderResult
 public;usePropertyDataProviderWithUnifiedSelection(props: PropertyDataProviderWithUnifiedSelectionProps): UsePropertyDataProviderWithUnifiedSelectionResult
 public;UsePropertyDataProviderWithUnifiedSelectionResult
