+++ common/changes/@itwin/components-react/presentation-filtering_docs_2023-01-24-08-43.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/components-react",
+      "comment": "API promotion",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/components-react"
+}
