+++ common/changes/@itwin/presentation-components/presentation-filtering_docs_2023-01-24-08-43.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/presentation-components",
+      "comment": "API promotion",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/presentation-components"
+}
