+++ presentation/components/src/presentation-components/instance-filter-builder/InstanceFilterBuilder.tsx
@@ -6,34 +6,47 @@
  * @module InstancesFilter
  */
 
+import "./InstanceFilterBuilder.scss";
 import * as React from "react";
 import { ActionMeta } from "react-select";
 import { PropertyDescription } from "@itwin/appui-abstract";
-import { PropertyFilter, PropertyFilterBuilder } from "@itwin/components-react";
-import { ClassInfo } from "@itwin/presentation-common";
+import { PropertyFilter, PropertyFilterBuilder, PropertyFilterBuilderProps } from "@itwin/components-react";
+import { assert } from "@itwin/core-bentley";
+import { IModelConnection } from "@itwin/core-frontend";
+import { ClassInfo, Descriptor } from "@itwin/presentation-common";
 import { translate } from "../common/Utils";
+import { NavigationPropertyEditorContextProps } from "../properties/NavigationPropertyEditor";
+import { getIModelMetadataProvider } from "./ECMetadataProvider";
 import { MultiTagSelect } from "./MultiTagSelect";
-import "./InstanceFilterBuilder.scss";
+import { PresentationInstanceFilterProperty } from "./PresentationInstanceFilterProperty";
+import { createInstanceFilterPropertyInfos, getInstanceFilterFieldName, InstanceFilterPropertyInfo } from "./Utils";
 
-/** @alpha */
-export interface InstanceFilterBuilderProps {
+/**
+ * Props for [[InstanceFilterBuilder]] component.
+ * @internal
+ */
+export interface InstanceFilterBuilderProps extends PropertyFilterBuilderProps {
+  /** Currently selected classes. */
   selectedClasses: ClassInfo[];
+  /** List of all available classes. */
   classes: ClassInfo[];
-  properties: PropertyDescription[];
+  /** Callback that is invoked when filter is changed. */
   onFilterChanged: (filter?: PropertyFilter) => void;
+  /** Callback that is invoked when class is selected. */
   onClassSelected: (selectedClass: ClassInfo) => void;
+  /** Callback that is invoked when class is de-selected. */
   onClassDeselected: (selectedClass: ClassInfo) => void;
+  /** Callback that is invoked when all selected classes are cleared. */
   onClearClasses: () => void;
-  onPropertySelected?: (property: PropertyDescription) => void;
-  ruleGroupDepthLimit?: number;
-  propertyRenderer?: (name: string) => React.ReactNode;
-  isDisabled?: boolean;
-  initialFilter?: PropertyFilter;
 }
 
-/** @alpha */
+/**
+ * Component for building complex instance filters based on instance properties. In addition to filter builder component
+ * it renders selector for classes that can be used to filter out available properties in filter rules.
+ * @internal
+ */
 export function InstanceFilterBuilder(props: InstanceFilterBuilderProps) {
-  const { selectedClasses, classes, properties, ruleGroupDepthLimit, onFilterChanged, onPropertySelected, onClassSelected, onClassDeselected, onClearClasses, isDisabled, initialFilter } = props;
+  const { selectedClasses, classes, onClassSelected, onClassDeselected, onClearClasses, ...restProps } = props;
 
   const onSelectChange = React.useCallback((_, action: ActionMeta<ClassInfo>) => {
     switch (action.action) {
@@ -65,19 +78,193 @@ export function InstanceFilterBuilder(props: InstanceFilterBuilderProps) {
         hideSelectedOptions={false}
         closeMenuOnSelect={false}
         isClearable={true}
-        isDisabled={isDisabled}
+        isDisabled={restProps.isDisabled}
       />
     </div>
     <div className="presentation-property-filter-builder">
       <PropertyFilterBuilder
-        properties={properties}
-        onFilterChanged={onFilterChanged}
-        onRulePropertySelected={onPropertySelected}
-        ruleGroupDepthLimit={ruleGroupDepthLimit}
-        propertyRenderer={props.propertyRenderer}
-        disablePropertySelection={isDisabled}
-        initialFilter={initialFilter}
+        {...restProps}
       />
     </div>
   </div>;
+}
+
+/**
+ * Custom hook that extracts properties and classes from [Descriptor]($presentation-common) and creates props that can be used by [[InstanceFilterBuilder]] component.
+ *
+ * This hook also makes sure that when classes are selected available properties list is updated to contain only properties found on selected classes and vice versa -
+ * when property is selected in one of the rules selected classes list is updated to contain only classes that has access to that property.
+ * @beta
+ */
+export function usePresentationInstanceFilteringProps(
+  descriptor: Descriptor,
+  imodel: IModelConnection,
+  initialClasses?: ClassInfo[]
+): Required<Pick<InstanceFilterBuilderProps, "properties" | "classes" | "selectedClasses" | "onClassSelected" | "onClassDeselected" | "onClearClasses" | "propertyRenderer" | "onRulePropertySelected" | "isDisabled">> {
+  const propertyInfos = React.useMemo(() => createInstanceFilterPropertyInfos(descriptor), [descriptor]);
+  const classes = React.useMemo((): ClassInfo[] => {
+    const uniqueClasses = new Map();
+    descriptor.selectClasses.forEach((selectClass) => uniqueClasses.set(selectClass.selectClassInfo.id, selectClass.selectClassInfo));
+    return [...uniqueClasses.values()];
+  }, [descriptor]);
+
+  const {
+    selectedClasses, onClassSelected, onClassDeselected, onClearClasses, isFilteringClasses, filterClassesByProperty,
+  } = useSelectedClasses(classes, imodel, initialClasses);
+  const { properties, isFilteringProperties } = useProperties(propertyInfos, selectedClasses, imodel);
+
+  const onRulePropertySelected = React.useCallback((property: PropertyDescription) => {
+    const propertyInfo = propertyInfos.find((info) => info.propertyDescription.name === property.name);
+    if (propertyInfo)
+      filterClassesByProperty(propertyInfo);
+  }, [propertyInfos, filterClassesByProperty]);
+
+  const propertyRenderer = React.useCallback((name: string) => {
+    const instanceFilterPropertyInfo = propertyInfos.find((info) => info.propertyDescription.name === name);
+    assert(instanceFilterPropertyInfo !== undefined);
+    return (
+      <PresentationInstanceFilterProperty
+        propertyDescription={instanceFilterPropertyInfo.propertyDescription}
+        fullClassName={instanceFilterPropertyInfo.className}
+        categoryLabel={instanceFilterPropertyInfo.categoryLabel}
+      />
+    );
+  }, [propertyInfos]);
+
+  return {
+    onRulePropertySelected,
+    onClearClasses,
+    onClassDeselected,
+    onClassSelected,
+    propertyRenderer,
+    properties,
+    classes,
+    selectedClasses,
+    isDisabled: isFilteringClasses || isFilteringProperties,
+  };
+}
+
+function useProperties(propertyInfos: InstanceFilterPropertyInfo[], selectedClasses: ClassInfo[], imodel: IModelConnection) {
+  const [filteredProperties, setFilteredProperties] = React.useState<InstanceFilterPropertyInfo[] | undefined>();
+  const [isFilteringProperties, setIsFilteringProperties] = React.useState(false);
+  const properties = React.useMemo(
+    () => (filteredProperties ?? propertyInfos).map((info) => info.propertyDescription),
+    [propertyInfos, filteredProperties]
+  );
+
+  // filter properties by selected classes
+  React.useEffect(() => {
+    if (selectedClasses.length === 0) {
+      setFilteredProperties(undefined);
+      return;
+    }
+
+    setIsFilteringProperties(true);
+    let disposed = false;
+    void (async () => {
+      const newFilteredProperties = await computePropertiesByClasses(propertyInfos, selectedClasses, imodel);
+      // istanbul ignore else
+      if (!disposed) {
+        setFilteredProperties(newFilteredProperties);
+        setIsFilteringProperties(false);
+      }
+    })();
+    return () => { disposed = true; };
+  }, [propertyInfos, selectedClasses, imodel]);
+
+  return {
+    properties,
+    isFilteringProperties,
+  };
+}
+
+function useSelectedClasses(classes: ClassInfo[], imodel: IModelConnection, initialClasses?: ClassInfo[]) {
+  const [selectedClasses, setSelectedClasses] = React.useState<ClassInfo[]>(initialClasses ?? []);
+  const [isFilteringClasses, setIsFilteringClasses] = React.useState(false);
+  const disposedRef = React.useRef(false);
+  React.useEffect(() => () => { disposedRef.current = true; }, []);
+
+  const firstRender = React.useRef(true);
+  React.useEffect(() => {
+    if (!firstRender.current)
+      setSelectedClasses([]);
+    firstRender.current = false;
+  }, [classes]);
+
+  const onClassSelected = React.useCallback((info: ClassInfo) => {
+    setSelectedClasses((prevClasses) => [...prevClasses, info]);
+  }, []);
+
+  const onClassDeselected = React.useCallback((classInfo: ClassInfo) => {
+    setSelectedClasses((prevClasses) => prevClasses.filter((info) => info.id !== classInfo.id));
+  }, []);
+
+  const onClearClasses = React.useCallback(() => {
+    setSelectedClasses([]);
+  }, []);
+
+  const filterClassesByProperty = React.useCallback((property: InstanceFilterPropertyInfo) => {
+    setIsFilteringClasses(true);
+    void (async () => {
+      const newSelectedClasses = await computeClassesByProperty(selectedClasses.length === 0 ? classes : selectedClasses, property, imodel);
+      // istanbul ignore else
+      if (!disposedRef.current) {
+        setSelectedClasses(newSelectedClasses);
+        setIsFilteringClasses(false);
+      }
+    })();
+  }, [selectedClasses, classes, imodel]);
+
+  return {
+    selectedClasses,
+    isFilteringClasses,
+    onClassSelected,
+    onClassDeselected,
+    onClearClasses,
+    filterClassesByProperty,
+  };
+}
+
+/** @internal */
+export function useFilterBuilderNavigationPropertyEditorContext(imodel: IModelConnection, descriptor: Descriptor) {
+  return React.useMemo<NavigationPropertyEditorContextProps>(() => ({
+    imodel,
+    getNavigationPropertyInfo: async (property) => {
+      const field = descriptor.getFieldByName(getInstanceFilterFieldName(property));
+      if (!field || !field.isPropertiesField())
+        return undefined;
+
+      return field.properties[0].property.navigationPropertyInfo;
+    },
+  }), [imodel, descriptor]);
+}
+
+async function computePropertiesByClasses(properties: InstanceFilterPropertyInfo[], classes: ClassInfo[], imodel: IModelConnection): Promise<InstanceFilterPropertyInfo[] | undefined> {
+  const metadataProvider = getIModelMetadataProvider(imodel);
+  const ecClassInfos = await Promise.all(classes.map(async (info) => metadataProvider.getECClassInfo(info.id)));
+  const filteredProperties: InstanceFilterPropertyInfo[] = [];
+  for (const prop of properties) {
+    // property should be shown if all selected classes are derived from property source class
+    if (ecClassInfos.every((info) => info && info.isDerivedFrom(prop.sourceClassId)))
+      filteredProperties.push(prop);
+  }
+
+  return filteredProperties.length === properties.length ? undefined : filteredProperties;
+}
+
+async function computeClassesByProperty(classes: ClassInfo[], property: InstanceFilterPropertyInfo, imodel: IModelConnection): Promise<ClassInfo[]> {
+  const metadataProvider = getIModelMetadataProvider(imodel);
+  const propertyClass = await metadataProvider.getECClassInfo(property.sourceClassId);
+  // istanbul ignore next
+  if (!propertyClass)
+    return classes;
+
+  const classesWithProperty: ClassInfo[] = [];
+  for (const currentClass of classes) {
+    // add classes that are derived from property source class
+    if (propertyClass.isBaseOf(currentClass.id))
+      classesWithProperty.push(currentClass);
+  }
+
+  return classesWithProperty;
 }
