+++ presentation/components/src/presentation-components/instance-filter-builder/Types.ts
@@ -6,36 +6,45 @@
  * @module InstancesFilter
  */
 
-import { PrimitiveValue, PropertyDescription } from "@itwin/appui-abstract";
+import { PrimitiveValue } from "@itwin/appui-abstract";
 import { PropertyFilterRuleGroupOperator, PropertyFilterRuleOperator } from "@itwin/components-react";
-import { ClassId, PropertiesField } from "@itwin/presentation-common";
+import { PropertiesField } from "@itwin/presentation-common";
 
-/** @alpha */
+/**
+ * Type that describes instance filter based on [Descriptor]($presentation-common) fields. It can be
+ * one filter condition or group of filter conditions joined by logical operator.
+ * @beta
+ */
 export type PresentationInstanceFilter = PresentationInstanceFilterConditionGroup | PresentationInstanceFilterCondition;
 
-/** @alpha */
+/**
+ * Data structure that describes group of filter condition joined by logical operator.
+ * @beta
+ */
 export interface PresentationInstanceFilterConditionGroup {
+  /** Operator that should be used to join conditions. */
   operator: PropertyFilterRuleGroupOperator;
+  /** Conditions in this group. */
   conditions: PresentationInstanceFilter[];
 }
 
-/** @alpha */
+/**
+ * Data structure that describes single filter condition.
+ * @beta
+ */
 export interface PresentationInstanceFilterCondition {
-  operator: PropertyFilterRuleOperator;
+  /** [PropertiesField]($presentation-common) that contains property used in this condition. */
   field: PropertiesField;
+  /** Operator that should be used to compare property value. */
+  operator: PropertyFilterRuleOperator;
+  /** Value that property should be compared to. */
   value?: PrimitiveValue;
 }
 
-/** @alpha */
-export interface InstanceFilterPropertyInfo {
-  sourceClassId: ClassId;
-  field: PropertiesField;
-  propertyDescription: PropertyDescription;
-  className: string;
-  categoryLabel?: string;
-}
-
-/** @alpha */
+/**
+ * Function that checks if supplied [[PresentationInstanceFilter]] is [[PresentationInstanceFilterConditionGroup]].
+ * @beta
+ */
 export function isPresentationInstanceFilterConditionGroup(filter: PresentationInstanceFilter): filter is PresentationInstanceFilterConditionGroup {
   return (filter as any).conditions !== undefined;
 }
