+++ ui/components-react/src/components-react.ts
@@ -86,12 +86,6 @@ export * from "./components-react/properties/renderers/value/UrlPropertyValueRen
 export * from "./components-react/properties/renderers/value/WithContextStyle";
 
 export * from "./components-react/filter-builder/FilterBuilder";
-export * from "./components-react/filter-builder/FilterBuilderRule";
-export * from "./components-react/filter-builder/FilterBuilderRuleGroup";
-export * from "./components-react/filter-builder/FilterBuilderRuleOperator";
-export * from "./components-react/filter-builder/FilterBuilderRuleProperty";
-export * from "./components-react/filter-builder/FilterBuilderRuleValue";
-export * from "./components-react/filter-builder/FilterBuilderState";
 export * from "./components-react/filter-builder/Operators";
 export * from "./components-react/filter-builder/Types";
 
@@ -239,4 +233,8 @@ export * from "./components-react/tree/controlled/internal/SparseTree";
 /**
  * @docs-group-description TypeConverters
  * Classes for working with Type Converters.
+ */
+/**
+ * @docs-group-description PropertyFilterBuilder
+ * Classes and components for working with PropertyFilterBuilder.
  */
