+++ ui/components-react/src/components-react/filter-builder/FilterBuilderRuleGroup.tsx
@@ -2,35 +2,35 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+/** @packageDocumentation
+ * @module PropertyFilterBuilder
+ */
+
+import "./FilterBuilderRuleGroup.scss";
 import * as React from "react";
 import { SvgAdd, SvgDelete } from "@itwin/itwinui-icons-react";
 import { Button, IconButton, Select, SelectOption } from "@itwin/itwinui-react";
 import { UiComponents } from "../UiComponents";
-import { PropertyFilterBuilderContext } from "./FilterBuilder";
+import { ActiveRuleGroupContext, PropertyFilterBuilderContext } from "./FilterBuilderContext";
 import { PropertyFilterBuilderRuleRenderer } from "./FilterBuilderRule";
 import { isPropertyFilterBuilderRuleGroup, PropertyFilterBuilderRuleGroup, PropertyFilterBuilderRuleGroupItem } from "./FilterBuilderState";
 import { PropertyFilterRuleGroupOperator } from "./Operators";
-import "./FilterBuilderRuleGroup.scss";
 
-/** @alpha */
-export interface ActiveRuleGroupContextProps {
-  activeElement: HTMLElement | undefined;
-  onFocus: React.FocusEventHandler<HTMLElement>;
-  onBlur: React.FocusEventHandler<HTMLElement>;
-  onMouseOver: React.MouseEventHandler<HTMLElement>;
-  onMouseOut: React.MouseEventHandler<HTMLElement>;
-}
-
-/** @alpha */
-export const ActiveRuleGroupContext = React.createContext<ActiveRuleGroupContextProps>(null!);
-
-/** @alpha */
+/**
+ * Props for [[PropertyFilterBuilderRuleGroupRenderer]] component.
+ * @internal
+ */
 export interface PropertyFilterBuilderRuleGroupRendererProps {
+  /** Path from [[PropertyFilterBuilder]] root to this rule group. */
   path: string[];
+  /** Rule group to render. */
   group: PropertyFilterBuilderRuleGroup;
 }
 
-/** @alpha */
+/**
+ * Component that renders group of rules in [[PropertyFilterBuilder]] component.
+ * @internal
+ */
 export function PropertyFilterBuilderRuleGroupRenderer(props: PropertyFilterBuilderRuleGroupRendererProps) {
   const { path, group } = props;
   const { actions, ruleGroupDepthLimit } = React.useContext(PropertyFilterBuilderContext);
@@ -75,13 +75,21 @@ export function PropertyFilterBuilderRuleGroupRenderer(props: PropertyFilterBuil
   </div>;
 }
 
-/** @alpha */
+/**
+ * Props for [[PropertyFilterBuilderRuleGroupOperator]] component.
+ * @internal
+ */
 export interface PropertyFilterBuilderRuleGroupOperatorProps {
+  /** Currently selected operator. */
   operator: PropertyFilterRuleGroupOperator;
+  /** Callback that is invoked when selected operator changes. */
   onChange: (operator: PropertyFilterRuleGroupOperator) => void;
 }
 
-/** @alpha */
+/**
+ * Component that renders [[PropertyFilterBuilderRuleGroup]] operator selector.
+ * @internal
+ */
 export function PropertyFilterBuilderRuleGroupOperator(props: PropertyFilterBuilderRuleGroupOperatorProps) {
   const { operator, onChange } = props;
