+++ ui/components-react/src/components-react/filter-builder/FilterBuilderRuleProperty.tsx
@@ -2,21 +2,36 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+/** @packageDocumentation
+ * @module PropertyFilterBuilder
+ */
+
 import * as React from "react";
 import { PropertyDescription } from "@itwin/appui-abstract";
 import { ComboBox, MenuItem, SelectOption } from "@itwin/itwinui-react";
 import { UiComponents } from "../UiComponents";
 
-/** @alpha */
+/**
+ * Props for [[PropertyFilterBuilderRuleProperty]] component.
+ * @internal
+ */
 export interface PropertyFilterBuilderRulePropertyProps {
+  /** List of available properties. */
   properties: PropertyDescription[];
+  /** Currently selected property. */
   selectedProperty?: PropertyDescription;
+  /** Callback that is invoked when selected property changes. */
   onSelectedPropertyChanged: (property?: PropertyDescription) => void;
+  /** Custom renderer for property item inside selector. */
   propertyRenderer?: (name: string) => React.ReactNode;
+  /** Specifies whether selector should be disabled or not. */
   isDisabled?: boolean;
 }
 
-/** @alpha */
+/**
+ * Component that renders [[PropertyFilterBuilderRuleRenderer]] property selector.
+ * @internal
+ */
 export function PropertyFilterBuilderRuleProperty(props: PropertyFilterBuilderRulePropertyProps) {
   const { selectedProperty, properties, onSelectedPropertyChanged, propertyRenderer, isDisabled } = props;
