+++ ui/components-react/src/components-react/filter-builder/FilterBuilderState.ts
@@ -2,39 +2,68 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+/** @packageDocumentation
+ * @module PropertyFilterBuilder
+ */
+
 import { Draft, produce } from "immer";
 import * as React from "react";
-import { PropertyDescription, PropertyValue } from "@itwin/appui-abstract";
+import { PropertyDescription, PropertyValue, PropertyValueFormat } from "@itwin/appui-abstract";
 import { Guid } from "@itwin/core-bentley";
 import { isUnaryPropertyFilterOperator, PropertyFilterRuleGroupOperator, PropertyFilterRuleOperator } from "./Operators";
 import { isPropertyFilterRuleGroup, PropertyFilter, PropertyFilterRule } from "./Types";
 
-/** @alpha */
+/**
+ * Data structure that describes [[PropertyFilterBuilder]] component state.
+ * @internal
+ */
 export interface PropertyFilterBuilderState {
+  /** Root group of rules in [[PropertyFilterBuilder]] component. */
   rootGroup: PropertyFilterBuilderRuleGroup;
 }
 
-/** @alpha */
+/**
+ * Type that describes [[PropertyFilterBuilder]] component group item.
+ * @internal
+ */
 export type PropertyFilterBuilderRuleGroupItem = PropertyFilterBuilderRuleGroup | PropertyFilterBuilderRule;
 
-/** @alpha */
+/**
+ * Data structure that describes [[PropertyFilterBuilder]] component rule group.
+ * @internal
+ */
 export interface PropertyFilterBuilderRuleGroup {
+  /** Id of this rule group. */
   id: string;
+  /** Id of rule group that this group is nested in. */
   groupId?: string;
+  /** Operator that should join items in this group. */
   operator: PropertyFilterRuleGroupOperator;
+  /** Items in this group. */
   items: PropertyFilterBuilderRuleGroupItem[];
 }
 
-/** @alpha */
+/**
+ * Data structure that describes [[PropertyFilterBuilder]] component single rule.
+ * @internal
+ */
 export interface PropertyFilterBuilderRule {
+  /** Id of this rule. */
   id: string;
+  /** Id of rule group that this rule is nested in. */
   groupId: string;
+  /** Property used in this rule. */
   property?: PropertyDescription;
+  /** Operator that should be used to compare property value. */
   operator?: PropertyFilterRuleOperator;
+  /** Value that property should be compared to. */
   value?: PropertyValue;
 }
 
-/** @alpha */
+/**
+ * Actions for controlling [[PropertyFilterBuilder]] component state.
+ * @internal
+ */
 export class PropertyFilterBuilderActions {
   constructor(private setState: (setter: (prevState: PropertyFilterBuilderState) => PropertyFilterBuilderState) => void) { }
 
@@ -42,6 +71,7 @@ export class PropertyFilterBuilderActions {
     this.setState(produce(updater));
   }
 
+  /** Adds new rule or group of rules to the group specified by path. */
   public addItem(path: string[], itemType: "RULE_GROUP" | "RULE") {
     this.updateState((state) => {
       const parentGroup = findRuleGroup(state.rootGroup, path);
@@ -52,6 +82,7 @@ export class PropertyFilterBuilderActions {
     });
   }
 
+  /** Removes item specified by path. */
   public removeItem(path: string[]) {
     function removeItemFromGroup(state: Draft<PropertyFilterBuilderState>, pathToItem: string[]) {
       const pathToParent = pathToItem.slice(0, -1);
@@ -72,6 +103,7 @@ export class PropertyFilterBuilderActions {
     });
   }
 
+  /** Sets operator of rule group specified by the path. */
   public setRuleGroupOperator(path: string[], operator: PropertyFilterRuleGroupOperator) {
     this.updateState((state) => {
       const group = findRuleGroup(state.rootGroup, path);
@@ -81,6 +113,7 @@ export class PropertyFilterBuilderActions {
     });
   }
 
+  /** Sets property of rule specified by the path. */
   public setRuleProperty(path: string[], property?: PropertyDescription) {
     this.updateState((state) => {
       const rule = findRule(state.rootGroup, path);
@@ -91,6 +124,7 @@ export class PropertyFilterBuilderActions {
     });
   }
 
+  /** Sets operator of rule specified by the path. */
   public setRuleOperator(path: string[], operator: PropertyFilterRuleOperator) {
     this.updateState((state) => {
       const rule = findRule(state.rootGroup, path);
@@ -102,6 +136,7 @@ export class PropertyFilterBuilderActions {
     });
   }
 
+  /** Sets value of rule specified by the path. */
   public setRuleValue(path: string[], value: PropertyValue) {
     this.updateState((state) => {
       const rule = findRule(state.rootGroup, path);
@@ -112,12 +147,19 @@ export class PropertyFilterBuilderActions {
   }
 }
 
-/** @alpha */
+/**
+ * Function to check if supplied [[PropertyFilterBuilderRuleGroupItem]] is [[PropertyFilterBuilderRuleGroup]].
+ * @internal
+ */
 export function isPropertyFilterBuilderRuleGroup(item: PropertyFilterBuilderRuleGroupItem): item is PropertyFilterBuilderRuleGroup {
   return (item as any).items !== undefined;
 }
 
-/** @alpha */
+/**
+ * Custom hook that creates state for [[PropertyFilterBuilder]] component. It creates empty state or initializes
+ * state from supplied initial filter.
+ * @internal
+ */
 export function usePropertyFilterBuilderState(initialFilter?: PropertyFilter) {
   const [state, setState] = React.useState<PropertyFilterBuilderState>(
     () => initialFilter ? convertFilterToState(initialFilter) : { rootGroup: createEmptyRuleGroup() }
@@ -125,6 +167,45 @@ export function usePropertyFilterBuilderState(initialFilter?: PropertyFilter) {
   const [actions] = React.useState(() => new PropertyFilterBuilderActions(setState));
 
   return { state, actions };
+}
+
+/** @internal */
+export function buildPropertyFilter(groupItem: PropertyFilterBuilderRuleGroupItem): PropertyFilter | undefined {
+  if (isPropertyFilterBuilderRuleGroup(groupItem))
+    return buildPropertyFilterFromRuleGroup(groupItem);
+  return buildPropertyFilterFromRule(groupItem);
+}
+
+function buildPropertyFilterFromRuleGroup(rootGroup: PropertyFilterBuilderRuleGroup): PropertyFilter | undefined {
+  if (rootGroup.items.length === 0)
+    return undefined;
+
+  const rules = new Array<PropertyFilter>();
+  for (const item of rootGroup.items) {
+    const rule = buildPropertyFilter(item);
+    if (!rule)
+      return undefined;
+    rules.push(rule);
+  }
+
+  if (rules.length === 1)
+    return rules[0];
+
+  return {
+    operator: rootGroup.operator,
+    rules,
+  };
+}
+
+function buildPropertyFilterFromRule(rule: PropertyFilterBuilderRule): PropertyFilter | undefined {
+  const { property, operator, value } = rule;
+  if (!property || operator === undefined)
+    return undefined;
+
+  if (!isUnaryPropertyFilterOperator(operator) && (value === undefined || value.valueFormat !== PropertyValueFormat.Primitive || value.value === undefined))
+    return undefined;
+
+  return { property, operator, value };
 }
 
 function createEmptyRule(groupId: string): PropertyFilterBuilderRule {
