+++ ui/components-react/src/test/filter-builder/Common.tsx
@@ -2,15 +2,15 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+
 import * as React from "react";
 import sinon from "sinon";
 import { render } from "@testing-library/react";
 import {
-  PropertyFilterBuilderContext, PropertyFilterBuilderContextProps, PropertyFilterBuilderRuleRenderingContext,
-  PropertyFilterBuilderRuleRenderingContextProps,
-} from "../../components-react/filter-builder/FilterBuilder";
+  ActiveRuleGroupContext, ActiveRuleGroupContextProps, PropertyFilterBuilderContext, PropertyFilterBuilderContextProps,
+  PropertyFilterBuilderRuleRenderingContext, PropertyFilterBuilderRuleRenderingContextProps,
+} from "../../components-react/filter-builder/FilterBuilderContext";
 import { PropertyFilterBuilderActions } from "../../components-react/filter-builder/FilterBuilderState";
-import { ActiveRuleGroupContext, ActiveRuleGroupContextProps } from "../../components-react/filter-builder/FilterBuilderRuleGroup";
 
 /** @internal */
 export function renderWithContext(
@@ -30,7 +30,7 @@ export function renderWithContext(
     ruleOperatorRenderer: rendererContextProps.ruleOperatorRenderer,
     ruleValueRenderer: rendererContextProps.ruleValueRenderer,
     propertyRenderer: rendererContextProps.propertyRenderer,
-    disablePropertySelection: rendererContextProps.disablePropertySelection,
+    isDisabled: rendererContextProps.isDisabled,
   };
 
   const activeGroupContextValue: ActiveRuleGroupContextProps = {
