+++ ui/components-react/src/test/filter-builder/FilterBuilderRule.test.tsx
@@ -2,6 +2,7 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+
 import { expect } from "chai";
 import * as React from "react";
 import sinon from "sinon";
@@ -131,7 +132,7 @@ describe("PropertyFilterBuilderRuleRenderer", () => {
       const { container, queryByText } = renderWithContext(
         <PropertyFilterBuilderRuleRenderer {...defaultProps} />,
         { actions, properties: [defaultProperty] },
-        { disablePropertySelection: true }
+        { isDisabled: true }
       );
 
       // attempt to open property selector
