+++ ui/components-react/src/test/filter-builder/Operators.test.ts
@@ -2,6 +2,7 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+
 import chai, { expect } from "chai";
 import chaiSubset from "chai-subset";
 import {
@@ -12,18 +13,18 @@ chai.use(chaiSubset);
 
 describe("getPropertyFilterOperators", () => {
   it("returns operators by type", () => {
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "boolean"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "boolean" })).to.containSubset([
       PropertyFilterRuleOperator.IsTrue,
       PropertyFilterRuleOperator.IsFalse,
     ]);
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "string"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "string" })).to.containSubset([
       PropertyFilterRuleOperator.IsEqual,
       PropertyFilterRuleOperator.IsNotEqual,
       PropertyFilterRuleOperator.IsNull,
       PropertyFilterRuleOperator.IsNotNull,
       PropertyFilterRuleOperator.Like,
     ]);
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "int"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "int" })).to.containSubset([
       PropertyFilterRuleOperator.IsEqual,
       PropertyFilterRuleOperator.IsNotEqual,
       PropertyFilterRuleOperator.IsNull,
@@ -33,7 +34,7 @@ describe("getPropertyFilterOperators", () => {
       PropertyFilterRuleOperator.Less,
       PropertyFilterRuleOperator.LessOrEqual,
     ]);
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "double"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "double" })).to.containSubset([
       PropertyFilterRuleOperator.IsEqual,
       PropertyFilterRuleOperator.IsNotEqual,
       PropertyFilterRuleOperator.IsNull,
@@ -43,7 +44,7 @@ describe("getPropertyFilterOperators", () => {
       PropertyFilterRuleOperator.Less,
       PropertyFilterRuleOperator.LessOrEqual,
     ]);
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "dateTime"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "dateTime" })).to.containSubset([
       PropertyFilterRuleOperator.IsEqual,
       PropertyFilterRuleOperator.IsNotEqual,
       PropertyFilterRuleOperator.IsNull,
@@ -53,7 +54,7 @@ describe("getPropertyFilterOperators", () => {
       PropertyFilterRuleOperator.Less,
       PropertyFilterRuleOperator.LessOrEqual,
     ]);
-    expect(getPropertyFilterOperators({name: "prop", displayLabel: "Prop", typename: "otherType"})).to.containSubset([
+    expect(getPropertyFilterOperators({ name: "prop", displayLabel: "Prop", typename: "otherType" })).to.containSubset([
       PropertyFilterRuleOperator.IsEqual,
       PropertyFilterRuleOperator.IsNotEqual,
       PropertyFilterRuleOperator.IsNull,
